issue fix in colombia config
issue fix in colombia config

+++ src/pandemia/world/world_factory/colombia_world_factory.py
@@ -115,7 +115,7 @@ class ColombiaWorldFactory(WorldFactory):
 
         # This will be the matrix returned by the function
         baseline_agents_travelling_matrix =\
-            np.zeros((num_of_regions, num_of_regions), dtype=int)
+            np.zeros((num_of_regions, num_of_regions), dtype=float)
 
         # Calculate local travel, rescaled according to step size
         share_matrix = np.zeros((num_of_regions, num_of_regions), dtype=float)
